backend/requirements.txt
alembic==1.4.1
Click==7.0
Flask==1.1.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
